requirements.txt
torch
torchvision
matplotlib
fastapi
uvicorn
Pillow
python-multipart
pandas
scikit-learn
scipy
evidently
evidently[notebooks]